Update README.md
The project is about utilizing deep learning methods to identify skin cancer indicators and distinguish it from other skin related conditions. In my project I've used a CNN and ResNet model to detect melanoma across the images provided. THe objective is to ensure that the model is able to distinguish melanoma from other indicators such as. Kurtosis. 

The project is measured in terms of accuracy, precision and recall score. 
There are sitll big improvements needed to improve its predictability which could be accomplished by having a larger sample dataset for melanoma images, more hyperparameter tuning etc.